requirements.txt
trio-websocket==0.10.2
pydantic==1.10.7
pytest-trio==0.8.0
